Refactor-5 Move ansible specfic fuunctions to ansible
